refactor: `usePwa` hook for core layout
